nadal nie działa
dużo zmian typów zmiennych, ciągle przekracza zakres

+++ POD_Blum-Micali_Generator/POD_Blum-Micali_Generator/Model/SiG.cs
@@ -9,7 +9,7 @@ namespace POD_Blum_Micali_Generator.Model
     //Generator 0 1
     class SiG
     {
-        public int Sig(int xn, int p)
+        public int Sig(long xn, int p)
         {
             if (xn > ((p - 1)/2))
             {
